backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# Deepgram SDK for speech-to-text
deepgram-sdk==3.2.7

# Groq SDK for LLM
groq==0.4.2

# Additional dependencies
python-dotenv==1.0.0
python-multipart==0.0.6
